# Properly cache deobfMcMCP to improve performance