Currently, when a project is being built with the use of ForgeGradle-2.3-SNAPSHOT, it is checked whether the project has been setup with `setupDecompWorkspace`. Unless that happens to be the case, the task `deobfMcMCP` is delegated for execution, along with the tasks it depends on.

During the execution phase of the Gradle lifecycle, the mentioned tasks may be marked as `SKIPPED` if they have already been performed earlier and their results have been cached for the project. However, it seems that merely determining these facts at this point may take Gradle 10 seconds or more - scaling to quite a lot of time, and, presumably, processing power as well.

This pull request aims to fix the described issue by making ForgeGradle check for the existence of the output of the `deobfMcMCP` task, instead of the entire decompiled workspace, before the execution phase starts. I attach a small comparison of the practical performance of the two methods on the same machine, each starting with a clean build of a small project, followed by two cached ones:

* [Old](https://pastebin.com/raw/4zLEpcyU): Clean build **44s**, cached builds **13s** and **12s**, respectively, for a cached average of **12.5s**;
* [New](https://pastebin.com/raw/VKGFSQPe): Clean build **31s**, cached builds **3s** and **2s**, respectively, for a cached average of **2.5s**.

(The seemingly improved clean build times are probably incidental, perhaps dependent on my network speed or some Gradle daemon intrinsics, however the improvement that follows for the cached builds is fully reproducible on my system, for all projects that I have thus far tested.)