@illyohs 2019-12-30
It's been explained  multiple times to you in the discord why this isn't an option or ideal randomly hacking away tasks for your edge case isn't going to happen Gradle takes time to check if a task needs to be run we don't control that.