@newbthenewbd 2019-12-30
I would be thankful for a quote of the `multiple times`. To the best of my present knowledge, the only time that something to this extent has been recently said on the Forge Project's Discord channel was affirmed by me... And was referring to manually excluding `deobfMcMCP` with each cached `gradle build`. This pull request is the result of Yours truly taking a look in ForgeGradle's code as to why the issue actually occurred. Indeed, I could have not received prior negative feedback about this particular change - simply because I have not publicly shared any information about it prior to sending this very request...

If You also took the time to take a single look at the proposed code, You might've noticed that the _current_ version is, by Your definition, already a hack, and quite a bit hackier than what I propose - instead of deciding whether `deobfMcMCP` should be executed depending on whether that very `deobfMcMCP` has been executed yet and its output is available, it decides depending on the output of the much heavier `setupDecompWorkspace`. I am sad to conclude that this must have had not happened.

Unless, of course, I am somehow horribly wrong and the change breaks some functionality that I had not tested or might otherwise deteriorate user experience? If that is the case, I am naturally open to hear about it - for the first time.

Should I however get no objective review of the pull request, yet have it declined still, I will have to conclude that Minecraft Forge is not a serious place to attempt to contribute improvements to, for free.